--- a/.gitignore
+++ b/.gitignore
@@ -3,3 +3,4 @@ __pycache__
 *.pyc
 *.log
 *.trace
+*.egg-info